requirements.txt
openai
python-dotenv
fire
requests
redis